KULEDOCS-669 -
Just changed isMandatoryTransfersSubType and isNonMandatoryTransfersSubType methods in TransactionalDocumentRuleBase. Added constants to TransactionalDocumentRuleBaseConstants for parameters and exception handling. Modified tests to remove extra debugging statements.

I ran tests for TransferOfFundsDocumentRule and they all passed. I assume values are being returned correctly for all tests based-on this. If not, then there should be some edge cases added to TransactionalDocumentRuleTestBase. That in itself would be a separate issue.