KULEDOCS-860 -
Final code checkin for this. I have to test in IE to verify that PDF IE workaround exists and is working. I may make some template changes, but as far as I can see, code is finished. I made some final changes to CashReceiptCoverSheetServiceImpl.

PDFDocument was not being closed correctly. This caused some corruption in the PDF stream. Using page importing to a PdfWriter instead of using the PdfStamper for the entire thing. PdfStamper stamps a stream, and the PdfWriter imports pages from the stream to use as a template. This means there is actually two streams in memory. The ByteArrayOutputStream for the PdfStamper, and another for the PdfWriter. An in-memory stream might cause problems if the PDF gets very large. I'm not sure how large it would have to be. There are 67 check-lines per page. There would have to be a lot of checks to make this an issue (hundred thousand maybe.)