Year end encumbrance forward test is now 100% green.
Fixed ojb mapping for prior year account service. No primary key was defined and an indicator field in the bo was a String instead of a boolean as it should have been.
Rewrote rules for year end encumbrance closing process including the cobol code as comments for easy debugging.
Added a new constructor to OriginEntry to replace initializeEntry in (i think) the year end service.

Resolves KULGL-172 & KULGL-175.