KULCG-66
ValidationCompletionUtils was injected into DataDictionaryServiceImpl. This means that ValidationCompletionUtils is loaded through Spring. This was needed because ValidationCompletionUtils needs to use the PersistenceStructureService to figure out the Class of the elements in a Collection. DataDictionaryServiceImpl then passes the ValidationCompletionUtils into the various Definition classes used by the data dictionary.

A getter for ValidationCompletionUtils was added to SpringServiceLocator.

LookupDaoOjb.isWriteable() and ObjectUtils.getPropertyType() were made smart in terms of being aware of attributes that may be collections.

Tests were updated as needed.

All tests passed.