KULOWF-51 Added a title attribute to DisbursementVoucherPayeeDetail.xml business object.
KualiXmlRuleAttributeImpl.java
Moved some string definitions out of the body of the code and defined final static Strings.  Refactored the static initializer so that it calls a static method.  This allows the static method to be called from the test class to reload the hash map after data dicitionary changes have been made.  Added code to get the DocumentEntries from the data dicitionary in addtition to the BusinessObjectEntries.  Added code so that the target attribute value can be determined from the last term in the xpath or from the ruledata function, as necessary.  Refactored the code that used the jdom Element type and replaced it with code that uses only the w3c classes.  Added code to use only the value after the last '.' if the title attribute value is fully qualified.

Constants.java
Added a value for STAND_IN_BUSINESS_OBJECT_FOR_ATTRIBUTES  which defines the business object to try and find business object labels in if not found otherwise.

Created KualiXmlRuleAttributeImplTest
This file has two test.  One to go over all of the ruleAttributes defined in KualiXmlRuleAttributeImpl.xml and make sure that a label is found.  The second test confirms that the labels are found in the data dicitiony by changing the value in the data dictionary to nonsense programatically and then verifying the change.

KualiXmlRuleAttriburteImplTest.xml
This file is a modified export of the ruleAttributes from workflow.  All of the namespace references/info have been removed.