KFSMI-5668 review all fp/purap fields that feed PDP to ensure field lengths match.
All field sizes that needed to be increased to ensure the application would not error due to inconsistent sizes have been fixed. After discussions with dev leads, the changes were constrained to fixing only those fields that needed to be increased to handle feeds from fp and purap in the pdp area.