KFSCD-45 KFSCD-25 Added method override for getSections, added new method shouldIncludeField and added the parameter service to the AccountInquirable.java file.  This is for determining whether the source of funds field should or should not appear on the Account inquiry screen based on the value of the DISPLAY_SOURCE_OF_FUNDS_IND parameter.
KFSCD-45 KFSCD-25 Added code to the AccountInquirable.java and AccountDocumentPresentationController.java files to not display the Source Of Funds field if the parameter does not exist plus ensuring that intended functionality is still working based on the vlaue if the parameter is there. Updated the extend class in AccountInquirable.java from KfsInquirableImpl to KualiInquirableImpl as KfsInquirableImpl does not exist in Financials 6.0.2.