UAF-475 Requeue Stuck Doucments Batch Job
Tech Note:

I also added a new institutional-config.properties file. This file needs to be added to the tomcat run command as an external config. I am doing this with a new environment variable named $UA_KFS_INSTITUTIONAL_CONFIG_DIR.  This should be set on all development workstations and in the tomcat run command for the environments.