UAF-2307 latest commit to deal with some build errors.
UAF-2305 UAF-2307 Adding the definition for the new job and steps to the spring-gl.xml.

UAF-2305 UAF-2307 Removed code to close the input/output files in the transferContents method in BudgetAdjustmentMergeSortStep.java as it was causing issues when it merged the next file because the output file was closed.

UAF-2305 UAF-2307 Corrected spelling in the ojb-gl.xml and corrected the check if the balance type code indicates that the transaction is for the current budget in BudgetAdjustmentCashTransferServiceImpl.java

UAF-2307 Fix for STE on save of financialObjectTypeCode.

Changes to Spring and the service impl class.

 UAF-2307 Fix to the fix for STE on save of financialObjectTypeCode.

Changes to the service impl class.

UAF-2305 UAF 2307 Fix to Fix for Seq Number STE

Changes to the service impl class to fix the Seq Number STE and also
provide better reporting.

UAF-2305 UAF 2307 Fix to universityFiscalYear STE

Changes to the service impl class to fix the universityFiscalYear STE
and also added a step to the job to combine reports.

UAF-2305 UAF-2307 Added missing column for the financialSystemOriginationCode in the BudgetAdjustmentTransaction table.

UAF-2305 UAF-2307 Adding the data dictionary file for BudgetAdjustmentTransaction.

UAF-2305 UAF-2307 Fix for reporting

Changes to spring and the service impl class to fix report order and
aggregation.