UAF-3327 Cash Transfer of Funds on Interfaced Budget Adjustments
This step 2 of the mod is an enhancement to remove all logic associated
with budget adjustment transactions crossing income streams. This second
commit is to address a problem found in testing where cash transfer
transactions associated with a new non-existent account were being
rejected prior to the scrubber. All related failing transactions should
now be rejected and demerged in the scrubber.